refactor: Calculator 리팩토링
