+++ src/main/java/lotto/utils/Calculator.java
@@ -1,4 +1,8 @@
-package lotto.domain;
+package lotto.utils;
+
+import lotto.domain.Lotto;
+import lotto.domain.Payment;
+import lotto.domain.Result;
 
 import java.util.List;
 
@@ -7,20 +11,20 @@ public class Calculator {
 
     private final Result result;
 
-    public Calculator(){
+    public Calculator() {
         this.result = new Result();
     }
 
     public Result calculateResult(List<Lotto> lottos, List<Integer> winningNumber, int bonusNumber) {
-        for(int i=0; i<lottos.size(); i++) {
+        for (int i = 0; i < lottos.size(); i++) {
             List<Integer> lottoNumbers = lottos.get(i).getLottoNumbers();
             int correctCount = 0;
             boolean correctBonus = false;
-            for(int j=0; j<LOTTO_LENGTH; j++) {
-                if (lottoNumbers.contains(winningNumber.get(j))){
+            for (int j = 0; j < LOTTO_LENGTH; j++) {
+                if (lottoNumbers.contains(winningNumber.get(j))) {
                     correctCount++;
                 }
-                if(lottoNumbers.contains(bonusNumber)){
+                if (lottoNumbers.contains(bonusNumber)) {
                     correctBonus = true;
                 }
             }
@@ -30,34 +34,34 @@ public class Calculator {
         return result;
     }
 
-    private void calculateRank(int correctCount, boolean correctBonus){
-        if(correctCount == 6) {
+    private void calculateRank(int correctCount, boolean correctBonus) {
+        if (correctCount == 6) {
             result.countRank(1);
             result.addIncome(2000000000);
             return;
         }
-        if(correctCount == 5 && correctBonus) {
+        if (correctCount == 5 && correctBonus) {
             result.countRank(2);
             result.addIncome(30000000);
             return;
         }
-        if(correctCount == 5) {
+        if (correctCount == 5) {
             result.countRank(3);
             result.addIncome(1500000);
             return;
         }
-        if(correctCount == 4) {
+        if (correctCount == 4) {
             result.countRank(4);
             result.addIncome(50000);
             return;
         }
-        if(correctCount == 3) {
+        if (correctCount == 3) {
             result.countRank(5);
             result.addIncome(5000);
         }
     }
 
-    public double calculateIncomeRate(Result result, Payment payment){
+    public double calculateIncomeRate(Result result, Payment payment) {
         return (result.getIncome() / payment.getPayment()) * 100;
     }
 }
